refactor: private libraries is now public
