+++ app/build.gradle
@@ -4,7 +4,7 @@ buildscript {
     }
 
     dependencies {
-        classpath 'io.fabric.tools:gradle:1.24.1'
+        classpath 'io.fabric.tools:gradle:1.24.2'
     }
 }
 apply plugin: 'com.android.application'
@@ -18,12 +18,12 @@ repositories {
 def gitCommitCount = Integer.parseInt('git rev-list --count HEAD'.execute([], project.rootDir).text.trim())
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion "26.0.2"
+    compileSdkVersion rootProject.ext.targetSdkVersion
+    buildToolsVersion rootProject.ext.buildToolsVersion
     defaultConfig {
         applicationId "moe.shizuku.fcmformojo"
-        minSdkVersion 24
-        targetSdkVersion 26
+        minSdkVersion rootProject.ext.minSdkVersion
+        targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode gitCommitCount
         versionName "0.3.7"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
@@ -69,17 +69,17 @@ dependencies {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
     testImplementation 'junit:junit:4.12'
-    implementation 'com.android.support:support-compat:26.0.2'
-    implementation 'com.android.support:support-core-utils:26.0.2'
-    implementation 'com.android.support:support-core-ui:26.0.2'
-    implementation 'com.android.support:support-fragment:26.0.2'
-    implementation 'com.android.support:recyclerview-v7:26.0.2'
-    implementation 'com.android.support:cardview-v7:26.0.2'
-    implementation 'moe.shizuku.preference:preference:2.0.6'
-    implementation 'moe.shizuku.preference:preference-dialog-android:2.0.6'
-    implementation 'moe.shizuku.preference:preference-simplemenu:2.0.6'
-    implementation 'moe.shizuku.support:support-utils:2.0'
-    implementation 'moe.shizuku.support:recyclerview-utils:2.0'
+    implementation "com.android.support:support-compat:$supportLibraryVersion"
+    implementation "com.android.support:support-core-utils:$supportLibraryVersion"
+    implementation "com.android.support:support-core-ui:$supportLibraryVersion"
+    implementation "com.android.support:support-fragment:$supportLibraryVersion"
+    implementation "com.android.support:recyclerview-v7:$supportLibraryVersion"
+    implementation "com.android.support:cardview-v7:$supportLibraryVersion"
+    implementation "moe.shizuku.preference:preference:$preferenceLibraryVersion"
+    implementation "moe.shizuku.preference:preference-dialog-android:$preferenceLibraryVersion"
+    implementation "moe.shizuku.preference:preference-simplemenu:$preferenceLibraryVersion"
+    implementation 'moe.shizuku.support:support-utils:2.0.0'
+    implementation 'moe.shizuku.support:recyclerview-utils:2.0.2'
     implementation 'moe.shizuku.support:privileged-api:23'
     implementation 'com.google.code.gson:gson:2.8.2'
     implementation 'com.squareup.retrofit2:retrofit:2.3.0'
