+++ app/src/main/java/moe/shizuku/fcmformojo/viewholder/GroupWhitelistItemViewHolder.java
@@ -16,7 +16,7 @@ import io.reactivex.functions.Consumer;
 import io.reactivex.schedulers.Schedulers;
 import moe.shizuku.fcmformojo.R;
 import moe.shizuku.fcmformojo.model.Group;
-import moe.shizuku.utils.recyclerview.BaseViewHolder;
+import moe.shizuku.support.recyclerview.BaseViewHolder;
 
 /**
  * Created by rikka on 2017/8/28.
